Clean unused transaction code
