+++ src/visor/visor.go
@@ -13,11 +13,8 @@ package visor
 import (
 	"errors"
 	"fmt"
-	"sort"
 
 	"time"
-
-	"github.com/sirupsen/logrus"
 
 	"github.com/SkycoinProject/skycoin/src/cipher"
 	"github.com/SkycoinProject/skycoin/src/coin"
@@ -43,7 +40,7 @@ type Visor struct {
 	blockchain  Blockchainer
 	history     Historyer
 	wallets     *wallet.Service
-	txnModel    *transactionModel
+	txns        transactionsGetter
 }
 
 // New creates a Visor for managing the blockchain database
@@ -104,7 +101,7 @@ func New(c Config, db *dbutil.DB, wltServ *wallet.Service) (*Visor, error) {
 		return nil, err
 	}
 
-	txnModel := transactionModel{
+	txns := transactionModel{
 		history:     history,
 		unconfirmed: utp,
 		blockchain:  bc,
@@ -118,7 +115,7 @@ func New(c Config, db *dbutil.DB, wltServ *wallet.Service) (*Visor, error) {
 		unconfirmed: utp,
 		history:     history,
 		wallets:     wltServ,
-		txnModel:    &txnModel,
+		txns:        &txns,
 	}
 
 	return v, nil
@@ -864,22 +861,6 @@ func (vs *Visor) InjectUserTransactionTx(tx *dbutil.Tx, txn coin.Transaction) (b
 	return known, head, inputs, err
 }
 
-// GetTransactionsForAddress returns the Transactions whose unspents give coins to a cipher.Address.
-// This includes both confirmed and unconfirmed transactions.
-func (vs *Visor) GetTransactionsForAddress(a cipher.Address) ([]Transaction, error) {
-	var txns []Transaction
-
-	if err := vs.db.View("GetTransactionsForAddress", func(tx *dbutil.Tx) error {
-		var err error
-		txns, _, err = vs.getTransactionsForAddresses(tx, []cipher.Address{a}, nil)
-		return err
-	}); err != nil {
-		return nil, err
-	}
-
-	return txns, nil
-}
-
 // GetTransaction returns a Transaction by hash.
 func (vs *Visor) GetTransaction(txnHash cipher.SHA256) (*Transaction, error) {
 	var txn *Transaction
@@ -1009,15 +990,17 @@ type AddrsFilter struct {
 // Match implements the TxFilter interface, this actually won't be used, only the 'Addrs' member is used.
 func (af AddrsFilter) Match(tx *Transaction) bool { return true }
 
+// ConfirmedTxFilter filters transactions base on whether they are confirmed.
 type ConfirmedTxFilter struct {
 	Confirmed bool
 }
 
+// Match implements the TxFilter interface, this actually won't be used, only the value of 'Confirmed' is used.
 func (cf ConfirmedTxFilter) Match(tx *Transaction) bool {
 	return tx.Status.Confirmed == cf.Confirmed
 }
 
-// NewTxFilter collects the transaction whose 'Confirmed' status matchs the parameter passed in.
+// NewConfirmedTxFilter collects the transaction whose 'Confirmed' status matchs the parameter passed in.
 func NewConfirmedTxFilter(isConfirmed bool) TxFilter {
 	return ConfirmedTxFilter{Confirmed: isConfirmed}
 }
@@ -1029,7 +1012,7 @@ func (vs *Visor) GetTransactions(flts []TxFilter, order SortOrder, page *PageInd
 	var pages uint64
 	if err := vs.db.View("GetTransactions", func(tx *dbutil.Tx) error {
 		var err error
-		txns, pages, err = vs.txnModel.GetTransactions(tx, flts, order, page)
+		txns, pages, err = vs.txns.GetTransactions(tx, flts, order, page)
 		return err
 	}); err != nil {
 		return nil, 0, err
@@ -1045,7 +1028,7 @@ func (vs *Visor) GetTransactionsWithInputs(flts []TxFilter, order SortOrder, pag
 	var pages uint64
 	if err := vs.db.View("GetTransactionsWithInputs", func(tx *dbutil.Tx) error {
 		var err error
-		txns, pages, err = vs.txnModel.GetTransactions(tx, flts, order, page)
+		txns, pages, err = vs.txns.GetTransactions(tx, flts, order, page)
 		if err != nil {
 			return err
 		}
@@ -1074,186 +1057,6 @@ func (vs *Visor) GetTransactionsWithInputs(flts []TxFilter, order SortOrder, pag
 	}
 
 	return txns, inputs, pages, nil
-}
-
-func (vs *Visor) getTransactions(tx *dbutil.Tx, flts []TxFilter, page *PageIndex) ([]Transaction, uint64, error) {
-	var addrFlts []AddrsFilter
-	var otherFlts []TxFilter
-	// Splits the filters into AddrsFilter and other filters
-	for _, f := range flts {
-		switch v := f.(type) {
-		case AddrsFilter:
-			addrFlts = append(addrFlts, v)
-		default:
-			otherFlts = append(otherFlts, f)
-		}
-	}
-
-	// Accumulates all addresses in address filters
-	addrs := accumulateAddressInFilter(addrFlts)
-
-	// Traverses all transactions to do collection if there's no address filter.
-	if len(addrs) == 0 {
-		if page != nil {
-			return vs.txnModel.traverseTxns(tx, otherFlts, page)
-		}
-		txns, err := vs.traverseTxns(tx, otherFlts)
-		if err != nil {
-			return nil, 0, err
-		}
-		return txns, 1, nil
-	}
-
-	// Gets addresses related transactions
-	txns, pages, err := vs.getTransactionsForAddresses(tx, addrs, page)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	// Checks other filters
-	f := func(txn *Transaction, flts []TxFilter) bool {
-		for _, flt := range flts {
-			if !flt.Match(txn) {
-				return false
-			}
-		}
-
-		return true
-	}
-
-	var retTxns []Transaction
-	for _, txn := range txns {
-		if f(&txn, otherFlts) {
-			retTxns = append(retTxns, txn)
-		}
-	}
-
-	return retTxns, pages, nil
-}
-
-func accumulateAddressInFilter(afs []AddrsFilter) []cipher.Address {
-	// Accumulate all addresses in address filters
-	addrMap := make(map[cipher.Address]struct{})
-	var addrs []cipher.Address
-	for _, af := range afs {
-		for _, a := range af.Addrs {
-			if _, exist := addrMap[a]; exist {
-				continue
-			}
-			addrMap[a] = struct{}{}
-			addrs = append(addrs, a)
-		}
-	}
-	return addrs
-}
-
-// getTransactionsForAddresses returns all addresses related transactions with pagination.
-// Including both confirmed and unconfirmed transactions.
-func (vs *Visor) getTransactionsForAddresses(tx *dbutil.Tx, addrs []cipher.Address, page *PageIndex) ([]Transaction, uint64, error) {
-	// Get the head block seq, for calculating the txn status
-	return vs.txnModel.GetTransactionsForAddresses(tx, addrs, page)
-}
-
-// traverseTxns traverses transactions in historydb and unconfirmed tx pool in db,
-// returns transactions that can pass the filters.
-func (vs *Visor) traverseTxns(tx *dbutil.Tx, flts []TxFilter) ([]Transaction, error) {
-	// Get the head block seq, for calculating the tx status
-	headBkSeq, ok, err := vs.blockchain.HeadSeq(tx)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return nil, errors.New("No head block seq")
-	}
-
-	var txns []Transaction
-
-	if err := vs.history.ForEachTxn(tx, func(_ cipher.SHA256, hTxn *historydb.Transaction) error {
-		if headBkSeq < hTxn.BlockSeq {
-			err := errors.New("Transaction block sequence is less than the head block sequence")
-			logger.Critical().WithError(err).WithFields(logrus.Fields{
-				"headBkSeq":  headBkSeq,
-				"txBlockSeq": hTxn.BlockSeq,
-			}).Error()
-			return err
-		}
-
-		h := headBkSeq - hTxn.BlockSeq + 1
-
-		bk, err := vs.blockchain.GetSignedBlockBySeq(tx, hTxn.BlockSeq)
-		if err != nil {
-			return fmt.Errorf("get block of seq: %v failed: %v", hTxn.BlockSeq, err)
-		}
-
-		if bk == nil {
-			return fmt.Errorf("block of seq: %d doesn't exist", hTxn.BlockSeq)
-		}
-
-		txn := Transaction{
-			Transaction: hTxn.Txn,
-			Status:      NewConfirmedTransactionStatus(h, hTxn.BlockSeq),
-			Time:        bk.Time(),
-		}
-
-		// Checks filters
-		for _, f := range flts {
-			if !f.Match(&txn) {
-				return nil
-			}
-		}
-
-		txns = append(txns, txn)
-		return nil
-	}); err != nil {
-		return nil, err
-	}
-
-	txns = sortTxns(txns)
-
-	// Gets all unconfirmed transactions
-	unconfirmedTxns, err := vs.unconfirmed.GetFiltered(tx, func(txn UnconfirmedTransaction) bool {
-		return true
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	for _, ux := range unconfirmedTxns {
-		txn := Transaction{
-			Transaction: ux.Transaction,
-			Status:      NewUnconfirmedTransactionStatus(),
-			Time:        uint64(timeutil.NanoToTime(ux.Received).Unix()),
-		}
-
-		if matched := func(flts []TxFilter, txn Transaction) bool {
-			for _, f := range flts {
-				if !f.Match(&txn) {
-					return false
-				}
-			}
-			return true
-		}(flts, txn); matched {
-			txns = append(txns, txn)
-		}
-	}
-	return txns, nil
-}
-
-// Sort transactions by block seq, if equal then compare hash
-func sortTxns(txns []Transaction) []Transaction {
-	sort.Slice(txns, func(i, j int) bool {
-		if txns[i].Status.BlockSeq < txns[j].Status.BlockSeq {
-			return true
-		}
-
-		if txns[i].Status.BlockSeq > txns[j].Status.BlockSeq {
-			return false
-		}
-
-		// If transactions in the same block, compare the hash string
-		return txns[i].Transaction.Hash().Hex() < txns[j].Transaction.Hash().Hex()
-	})
-	return txns
 }
 
 // AddressBalances computes the total balance for cipher.Addresses and their coin.UxOuts
@@ -2132,80 +1935,80 @@ func (vs *Visor) AddressCount() (uint64, error) {
 }
 
 // GetVerboseTransactionsForAddress returns verbose transaction data for a given address
-func (vs *Visor) GetVerboseTransactionsForAddress(a cipher.Address) ([]Transaction, [][]TransactionInput, error) {
-	var txns []Transaction
-	var inputs [][]TransactionInput
+// func (vs *Visor) GetVerboseTransactionsForAddress(a cipher.Address) ([]Transaction, [][]TransactionInput, error) {
+// 	var txns []Transaction
+// 	var inputs [][]TransactionInput
 
-	if err := vs.db.View("GetVerboseTransactionsForAddress", func(tx *dbutil.Tx) error {
-		var err error
-		txns, _, err = vs.getTransactionsForAddresses(tx, []cipher.Address{a}, nil)
-		if err != nil {
-			logger.Errorf("GetVerboseTransactionsForAddress: vs.GetTransactionsForAddress failed: %v", err)
-			return err
-		}
+// 	if err := vs.db.View("GetVerboseTransactionsForAddress", func(tx *dbutil.Tx) error {
+// 		var err error
+// 		txns, _, err = vs.getTransactionsForAddresses(tx, []cipher.Address{a}, nil)
+// 		if err != nil {
+// 			logger.Errorf("GetVerboseTransactionsForAddress: vs.GetTransactionsForAddress failed: %v", err)
+// 			return err
+// 		}
 
-		if len(txns) == 0 {
-			return nil
-		}
+// 		if len(txns) == 0 {
+// 			return nil
+// 		}
 
-		head, err := vs.blockchain.Head(tx)
-		if err != nil {
-			logger.Errorf("GetVerboseTransactionsForAddress: vs.blockchain.Head failed: %v", err)
-			return err
-		}
+// 		head, err := vs.blockchain.Head(tx)
+// 		if err != nil {
+// 			logger.Errorf("GetVerboseTransactionsForAddress: vs.blockchain.Head failed: %v", err)
+// 			return err
+// 		}
 
-		inputs = make([][]TransactionInput, len(txns))
+// 		inputs = make([][]TransactionInput, len(txns))
 
-		for i, txn := range txns {
-			// If the txn is confirmed, use the time of the block previous
-			// to the block in which the transaction was executed,
-			// else use the head time for unconfirmed blocks.
-			t := head.Time()
-			if txn.Status.Confirmed && txn.Status.BlockSeq > 0 {
-				prevBlock, err := vs.blockchain.GetSignedBlockBySeq(tx, txn.Status.BlockSeq-1)
-				if err != nil {
-					return err
-				}
+// 		for i, txn := range txns {
+// 			// If the txn is confirmed, use the time of the block previous
+// 			// to the block in which the transaction was executed,
+// 			// else use the head time for unconfirmed blocks.
+// 			t := head.Time()
+// 			if txn.Status.Confirmed && txn.Status.BlockSeq > 0 {
+// 				prevBlock, err := vs.blockchain.GetSignedBlockBySeq(tx, txn.Status.BlockSeq-1)
+// 				if err != nil {
+// 					return err
+// 				}
 
-				if prevBlock == nil {
-					return fmt.Errorf("GetVerboseTransactionsForAddress prevBlock seq=%d missing", txn.Status.BlockSeq-1)
-				}
+// 				if prevBlock == nil {
+// 					return fmt.Errorf("GetVerboseTransactionsForAddress prevBlock seq=%d missing", txn.Status.BlockSeq-1)
+// 				}
 
-				t = prevBlock.Block.Head.Time
-			}
+// 				t = prevBlock.Block.Head.Time
+// 			}
 
-			txnInputs := make([]TransactionInput, len(txn.Transaction.In))
-			for j, inputID := range txn.Transaction.In {
-				uxOuts, err := vs.history.GetUxOuts(tx, []cipher.SHA256{inputID})
-				if err != nil {
-					logger.Errorf("GetVerboseTransactionsForAddress: vs.history.GetUxOuts failed: %v", err)
-					return err
-				}
-				if len(uxOuts) == 0 {
-					err := fmt.Errorf("uxout of %v does not exist in history db", inputID.Hex())
-					logger.Critical().Error(err)
-					return err
-				}
+// 			txnInputs := make([]TransactionInput, len(txn.Transaction.In))
+// 			for j, inputID := range txn.Transaction.In {
+// 				uxOuts, err := vs.history.GetUxOuts(tx, []cipher.SHA256{inputID})
+// 				if err != nil {
+// 					logger.Errorf("GetVerboseTransactionsForAddress: vs.history.GetUxOuts failed: %v", err)
+// 					return err
+// 				}
+// 				if len(uxOuts) == 0 {
+// 					err := fmt.Errorf("uxout of %v does not exist in history db", inputID.Hex())
+// 					logger.Critical().Error(err)
+// 					return err
+// 				}
 
-				input, err := NewTransactionInput(uxOuts[0].Out, t)
-				if err != nil {
-					logger.Errorf("GetVerboseTransactionsForAddress: NewTransactionInput failed: %v", err)
-					return err
-				}
+// 				input, err := NewTransactionInput(uxOuts[0].Out, t)
+// 				if err != nil {
+// 					logger.Errorf("GetVerboseTransactionsForAddress: NewTransactionInput failed: %v", err)
+// 					return err
+// 				}
 
-				txnInputs[j] = input
-			}
+// 				txnInputs[j] = input
+// 			}
 
-			inputs[i] = txnInputs
-		}
+// 			inputs[i] = txnInputs
+// 		}
 
-		return nil
-	}); err != nil {
-		return nil, nil, err
-	}
+// 		return nil
+// 	}); err != nil {
+// 		return nil, nil, err
+// 	}
 
-	return txns, inputs, nil
-}
+// 	return txns, inputs, nil
+// }
 
 // OutputsFilter used as optional arguments in GetUnspentOutputs method
 type OutputsFilter func(outputs coin.UxArray) coin.UxArray
